requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
anthropic==0.25.0
requests==2.31.0
httpx==0.24.1